+++ testing/adios2/engine/ssc/TestSscXgc3WayMatchedSteps.cpp
@@ -71,7 +71,7 @@ void coupler(const Dims &shape, const Dims &start, const Dims &count,
     adios2::Engine g_to_c_engine = g_to_c_io.Open("g_to_c", adios2::Mode::Read);
     g_to_c_engine.LockReaderSelections();
 
-    for (int i = 0; i < steps; ++i)
+    for (size_t i = 0; i < steps; ++i)
     {
         x_to_c_engine.BeginStep();
         auto x_to_c_var = x_to_c_io.InquireVariable<float>("x_to_c");
@@ -142,7 +142,7 @@ void xgc(const Dims &shape, const Dims &start, const Dims &count,
     adios2::Engine c_to_x_engine = c_to_x_io.Open("c_to_x", adios2::Mode::Read);
     c_to_x_engine.LockReaderSelections();
 
-    for (int i = 0; i < steps; ++i)
+    for (size_t i = 0; i < steps; ++i)
     {
         x_to_c_engine.BeginStep();
         GenData(x_to_c_data, i, start, count, shape);
@@ -193,7 +193,7 @@ void gene(const Dims &shape, const Dims &start, const Dims &count,
     std::vector<float> c_to_g_data;
     std::vector<float> g_to_c_data(datasize);
 
-    for (int i = 0; i < steps; ++i)
+    for (size_t i = 0; i < steps; ++i)
     {
         c_to_g_engine.BeginStep(StepMode::Read, 5);
         auto c_to_g_var = c_to_g_io.InquireVariable<float>("c_to_g");
